pyproject.toml
[project]
name = "mor-slm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=0.20.0",
    "datasets>=2.12.0",
    "einops>=0.6.0",
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.0.0",
    "seaborn>=0.11.0",
    "tokenizers>=0.13.0",
    "torch>=2.0.0",
    "tqdm>=4.65.0",
    "transformers>=4.30.0",
    "wandb>=0.15.0",
]
